requirements.txt
Flask

Flask-Mail
Flask-Login
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Dance

requests

mysqlclient

PyMySQL

pdfkit

numpy

sympy

pytest

WTForms

gunicorn